refactor code to share across different devices
[ghstack-poisoned]
