+++ torch/cuda/__init__.py
@@ -25,8 +25,8 @@ import torch
 import torch._C
 from torch.types import Device
 from .. import device as _device
-from .._utils import classproperty
-from ._utils import _dummy_type, _get_device_index
+from .._utils import classproperty, _dummy_type
+from ._utils import _get_device_index
 from .graphs import (
     CUDAGraph,
     graph,
