+++ torch/cuda/graphs.py
@@ -3,7 +3,7 @@ from typing import Optional
 
 import torch
 from torch.utils import _pytree
-from ._utils import _dummy_type
+from .._utils import _dummy_type
 
 if not hasattr(torch._C, "_CudaStreamBase"):
     # Define dummy base classes
