+++ torch/xpu/__init__.py
@@ -13,7 +13,8 @@ from typing import Any, Callable, Dict, List, Optional, Tuple, Union
 import torch
 import torch._C
 from .. import device as _device
-from ._utils import _dummy_type, _get_device_index
+from .._utils import _dummy_type
+from ._utils import _get_device_index
 from .streams import Event, Stream
 
 _initialized = False
