+++ torch/xpu/_utils.py
@@ -37,19 +37,3 @@ def _get_device_index(
         if isinstance(device, torch.xpu.device):
             return device.idx
     return _torch_get_device_index(device, optional, allow_cpu)
-
-
-def _dummy_type(name: str) -> type:
-    def get_err_fn(is_init: bool):
-        def err_fn(obj, *args, **kwargs):
-            if is_init:
-                class_name = obj.__class__.__name__
-            else:
-                class_name = obj.__name__
-            raise RuntimeError(f"Tried to instantiate dummy base class {class_name}")
-
-        return err_fn
-
-    return type(
-        name, (object,), {"__init__": get_err_fn(True), "__new__": get_err_fn(False)}
-    )
