+++ torch/xpu/streams.py
@@ -2,7 +2,7 @@ import ctypes
 
 import torch
 from torch._streambase import _EventBase, _StreamBase
-from ._utils import _dummy_type
+from .._utils import _dummy_type
 
 
 if not hasattr(torch._C, "_XpuStreamBase"):
